Update Build Status pages for new branches. (#36)
This update aligns with the addition of branches `release/1.5` and `release/1.6`. Since those branches take over from the current state of `release/2.0` and `master`, respectively, their pages are clones of the corresponding pages with the URLs updated.

As per the change to branch layout and milestones, `release/2.0` has been removed from the Sidebar.